Added svg template decorator. Fixes #62
Added template support for diagram designer. WIP #62

WIP #62 Added svgs

Added svgs for #62

These are a good starting point but they have a lot of room for
improvement. For now, blue are main layer types, octogons are loss
layers, and grey are utility. Labels and dropout layers are yellow
and input layers are also octogorns.